requirements.txt
requests~=2.32.3
synology-api~=0.7.3
python-dotenv~=1.0.1
tqdm~=4.66.6
tqdm-loggable~=0.2
boto3~=1.35.55
botocore~=1.35.55